backend/requirements.txt
fastapi
uvicorn
httpx
groq
anthropic
python-dotenv
